Добавлен Deck класс + бд

+++ src/sources/handsdeck.cpp
@@ -18,7 +18,7 @@
       
     void handsdeck::print()
     {
-        for (int i = 0;i < deck.size();i++)
+        for (size_t i = 0; i < deck.size(); i++)
         {
             std::cout << deck[i].getFigure() << ' ';
         }
